Corrected format error - had listed setPrice in Product as an int.

+++ Shopping_Basket/app/src/main/java/com/codeclan/shopping_basket/Product.java
@@ -29,7 +29,7 @@ public class Product {
         return this.price;
     }
 
-    public void setPrice(int newPrice){
+    public void setPrice(double newPrice){
         this.price = newPrice;
     }
